+++ packages/terra-form-select/src/search/Frame.jsx
@@ -145,6 +145,8 @@ class Frame extends React.Component {
       isPositioned: false,
       hasSearchChanged: false,
       searchValue: '',
+      isAbove: false,
+      resizeOnSearch: false,
     };
 
     this.ariaLabel = this.ariaLabel.bind(this);
@@ -174,7 +176,7 @@ class Frame extends React.Component {
   }
 
   componentDidUpdate(previousProps, previousState) {
-    if (FrameUtil.shouldPositionDropdown(previousState, this.state, this.dropdown)) {
+    if (FrameUtil.shouldPositionDropdown(previousState, this.state, this.dropdown, this.props.isTouchAccessible)) {
       clearTimeout(this.debounceTimer);
       this.debounceTimer = setTimeout(this.positionDropdown, !previousState.isOpen ? 0 : 100);
     }
@@ -287,7 +289,7 @@ class Frame extends React.Component {
 
     const { dropdownAttrs, maxHeight, isTouchAccessible } = this.props;
 
-    this.setState(FrameUtil.dropdownPosition(dropdownAttrs, this.select, this.dropdown, maxHeight, isTouchAccessible));
+    this.setState(prevState => (FrameUtil.dropdownPosition(dropdownAttrs, this.select, this.dropdown, maxHeight, prevState, isTouchAccessible)));
   }
 
   /**
@@ -424,6 +426,7 @@ class Frame extends React.Component {
       isOpen: true,
       hasSearchChanged: true,
       searchValue,
+      resizeOnSearch: true,
     });
 
     if (this.props.onSearch) {