+++ packages/terra-form-select/tests/jest/FrameUtil.test.jsx
@@ -43,6 +43,8 @@ describe('FrameUtil', () => {
         maxHeight: 100,
         isAbove: false,
         isPositioned: true,
+        resizeOnSearch: false,
+        top: undefined,
       };
 
       expect(FrameUtil.dropdownPosition(props, target, dropdown)).toEqual(expected);